Consolidate context-parallel summaries and routing
Move summary algebra and its canonical math/terminology walkthrough into
attn_gym.linear.context_parallel, alongside ownership plans and the collective
recipe. Delete state_summary.py without a forwarding shim and update every
in-repo import and documentation reference. The public helpers now import from
attn_gym.linear.context_parallel: neutral_summary, merge_state, compose_summaries.

Inline the small _exact_matmul wrapper. Decorate merge_state and compose_summaries
with the existing IEEE-FP32 context manager, keeping casts and matrix products
visible while restoring the caller's precision setting on return or exception.
Preserve the recently landed FP32 folding and device work-item scheduling; no
kernel or scheduler algorithms change.

Validation:
- CPU ownership/algebra suite: 92 passed without CuTeDSL installed.
- H100 staged-summary/algebra/simulated-CP checks: 34 passed, 13 Mega cases skipped
  because they require SM100/SM103.
- Two-H100 public KDA/GDN NCCL suite: 14 passed, including changing-layout CUDA
  Graph replay, final-state/output-only gradients and convolution halo gradients.
- Enhanced precision test checks IEEE-FP32 results with a TF32 caller setting,
  and precision restoration after both successful calls and shape errors.
- Scoped prek/ruff/diff checks passed. A clean-context simplifier review found no
  required fixes in the consolidation; other runtime definitions are unchanged.

GPU environment: Python 3.13.13, torch 2.15.0.dev20260904+cu132, CuTeDSL 4.7.1.

stack-info: PR: https://github.com/meta-pytorch/attention-gym/pull/500, branch: drisspg/stack/139
